+++ Sources/SwiftUIKit/Views/CustomRoundedRectangle.swift
@@ -59,7 +59,7 @@ public struct CustomRoundedRectangle: Shape {
     }
 }
 
-struct CustomRoundedRectangle_Previewsectangle: PreviewProvider {
+struct CustomRoundedRectangle_Previews: PreviewProvider {
     static var previews: some View {
         CustomRoundedRectangle(topLeft: 10, topRight: 20, bottomLeft: 30, bottomRight: 40)
             .foregroundColor(.red)